Display players as players intead of blaze
Send the add player packet instead of the add mob packet with the
blaze entity type ID. We can do this now that we also have a tab list
(players must be in the tab list for clients to render them). Also
update our entity rotation stuff to head and body specific rotations
(only for players at the moment).

There is currently the very spurious bug that a player's chunk cache
radius and centre get set to some seemingly random value on join. I
think this happens because those fields are right behind the send
buffer and some of the sending code (2 places) doesn't bounds check
and can therefore write outside the send buffer. We now bounds check
all writes to the send buffer.