Simplify packet sending and receiving
Make enums for clientbound and serverbound packet types, so we don't
need to go through all the code and update packet IDs manually when
updating to a new Minecraft version.

Move the serverbound packet processing code to a separate function to
get rid of the deep indentation and to make it easier to see what the
player brain tick function does.

Get rid of all the manual packet size calculations and determine the
packet size based on the amount of written packet data instead.
Manually calculating the packet size is very error prone, hard to
debug and requires a lot of maintaince when a packet type's format
changes.