Merge entities and brains
Currently the distinction between brains and entities is useless and
only results in annoyance: data that needs to be accessed by the same
code is stored in different places.

The main motivation for brains was to separate entity data other
entities need to access (such as players when sending packets) and
data only the brain of the entity needs to know about. This is useful
for plugins that provide their own brains, and in case multiple
entities are controlled by a single brain. However, we're currently
far from supporting of plugins, and Blaze won't include brains that
control multiple entities (since vanilla doesn't have these).

Perhaps in the future brains will be re-introduced.