Better handling of packet errors
Ensure we always end timed blocks. At some point we'll need to send
kick messages instead of disconnecting the client abruptly. Ideally we
would want to trash all current outgoing packets, write the disconnect
packet, and terminate the connection. But then we need to track the
current outgoing packet size, make sure the client is not still in the
LOGIN state, and perhaps other things need to be taken care of.