Basic block update propagation
The big change in this commit is the addition of block updates! This
implements the simplest block updates in the game and propagates these
updates. A block update cycle is started whenever a player places or
breaks a block.

I also implemented a few additional item-use-on's and moved all the
place function parameters into a context structure. This allows us to
change parameters without having to update all arguments.