Streamline entity tracking logic
The entity tracking logic is a bit hard to follow, so I streamlined it
a bit and added comments for further clarification. This fixes a bug
where entities aren't retracked when they move out of range and back
into range; we incorrectly started tracking an entity after deciding
to remove it, but before actually removing it.