Fix build script linker errors
Some people have reported that the build script doesn't work; they get
a bunch of linker errors. Switching the order of the libraries and
source files, seems to fix it.

I'm not entirely sure what's going on. According to the GCC docs, the
order in the command line specifies the order in which object files
are loaded and libraries are checked for references. However, the
build script works fine on my Raspberry Pi with GCC. Clang also
doesn't seem to exhibit this issue, at least on my Mac.