Broken redstone dust power propagation
Exploratory code for redstone dust power propagation. Levers and
redstone torches can power redstone dust, and power will propagate
through a redstone line.

The system in this commit is very messy and will likely introduce tons
of more messy code. It is also doomed to be very slow; see the
comments. We should switch over to a better system while it's still
easy to do.