Update to 1.18.1
The major changes in this update are the increased world height, the
merging of the light update packet and the chunk data packet with
additional changes to the network chunk format, and the modified chunk
format on disk.

This update also introduces a new NBT parser with a much nicer API.
The old parser was parsing some things incorrectly and I had another
NBT parser lying around, so I thought I might as well switch over to
the other one. There is a slight performance regression due to this,
which should be investigated at some point.

The major issue with the increased world height is that the game also
has negative Y values now, which means section indexing needs extra
work and some types that were previously unsigned become signed.