Refactor block models
The code currently assumes that block collision models and block
support models are separate concepts: the collision model is only used
for collisions and the support model only for supporting. However,
this is not true in vanilla for some dumb reason. For example, sea
pickles supporting uses collision models instead of support models.
This is bad for us, because we don't have efficient lookup structures
for support data of collision models.

To account for this, I merged collision models and support models into
1 struct: the block model. A block model has both collision data and
support data, so it can be used in both circumstances. A block state
can have a separate collision block model and support block model like
vanilla.

Maybe we'll separate the concepts again in the future when I come up
with a better work around. Maybe some extra flags/data on the support
model will work. We need to allow e.g. sea pickles to be placed on
leaves (empty support model).