Beginnings of multi-world support
Switch from BlockPos to WorldBlockPos, which includes a world ID along
with a BlockPos. As an alternative, I also tried adding a separate
world ID parameter to all functions, but that's a lot more cumbersome.
Easier to just replace BlockPos by WorldBlockPos everywhere.

The only downside is that some functions acting on a BlockPos need to
be duplicated for WorldBlockPos (currently only WorldBlockPosRel had
to be added).

Some other thoughts about WorldBlockPos: it's not very friendly for
bulk storing block positions in the same world. Though BlockPos isn't
either, so that's not a real concern.

Another concern could be that we're duplicating information, e.g. if a
player uses a block we have the player's world and the block's world,
though they're always the same. Some care might need to be taken to
handle such cases properly in future API functions.