Rework chunk lookup system
Previously we stored chunks in buckets, where each bucket corresponds
to 1 hash code of the chunk position. This was very simple to
implement, but we moved chunks around to refill empty spots after
deleting other chunks. This means chunk pointers were not stable. This
will likely be annoying if we do async chunk loading. Switch to a hash
table that maps chunk positions to indices into a separate array of
chunks.