Avoid using bit-fields and update Tracy
Bit-fields have awful guarantees in the C specification. As far as I
can tell, there's no guarantee that a u64 bit-field will even be put
in a u64. We also had a bug where the last 8 bits of a
PackedWorldChunkPos could be filled with random junk (I believe that
was it anyway) when initialising it. This resulted in failed
PackedWorldChunkPos comparisons and incorrect hashes, which in turn
resulted in failed chunk lookups.

This commit also updates Tracy from 0.7.8 to 0.8.2. It has a bunch of
extra things on modern Linux.