Start avoiding block types
I want to move to a block behaviour system in which block types aren't
explicitly used. Instead, I want to look up properties of block states
and implement behaviours based on that. All properties will still be
set up using block types in blockinit.c though.

The major downside of the current codebase is that we have all kinds
of block property getters (e.g. getting the water level) that may need
to be edited whenever a new block is added. This is incredibly
cumbersome. It would be more convenient if we could add a block type
together with all its properties to the initialiser.

Another approach would be to emulate block type classes (as in Java).
The pro here is that we can replicate the Mojang class layout, which
would allow us to synchronise changes to our codebase more easily. The
problem is that it introduces code path duplication (e.g. waterlogged
blocks have similar behaviours and we need to add code for shared
behaviour to each block type separately).