Fix bugs in the chunk hash map
Currently resizing seems to be broken. It's done in-place by enlarging
the entries array and then rehashing every element. I don't think this
always yields a properly formatted hash map; entries may be in
unexpected places.

Instead, we malloc the entries array and move all entries over to a
fresh malloc when growing the entries array.

In line with the previous commit, we also malloc the individual chunk
holder structs now. I expect this will change back in the future.