Simplify chunk loader
Currently the chunk loader keeps track of chunks that need to be
loaded, chunks that need to be unloaded and chunks that have been
loaded, in separate lists. It's easier to track all of those in a
single list of chunks that need to be updated. We do this now. For
every chunk that needs to be updated, we call an update function that
sorts out the details.