Separate packet parsing and processing
In the future we will probably do async packet reading, i.e. reading
from the OS socket buffer, decrypting, decompressing and validating
packet sizes. The goal of this commit is to prepare us for that.

Just reading from the OS buffer seems to take 2+ microseconds, which
is significant if we have 1000 players online. I imagine the packet
processing itself takes a lot less time, but we'll have to see. It's
probably worthwhile to do decryption and decompression async anyhow.

I'm not sure yet how async we should be going with this. My current
plan is to collect all inbound packets at the start of each tick in a
multi-threaded for-loop. But doing packet reading fully async (+ some
processing?) is also a possibility. The latter has the benefit that we
can e.g. respond to server list ping requests immediately, which
results in a more accurate latency number in the server list.