Fix chunk flags and wait for 3x3 to load
Due to a recent patch, both the main thread and the IO threads were
reading and modifying the statusFlags field, which is no bueno. We now
only read/write flags from the main thread.

Additionally, the 3x3 of chunks around a chunk with interest are
loaded now. We run the lighting algorithm on all of them, and only
when all the 3x3 chunks have been lit, do we allow the centre chunk to
become a "ready" chunk. That is, a chunk that the rest of the server
can interact with.