Experiment by allowing certain palette sizes only
The performance drop after implementing palettes completely was 8-17%
in the light engine across various types of worlds. Can't say I like
that. Not to mention that I feel the light engine is still quite slow:
during my testing it's far from hitting my single-threaded memory
bandwidth. To experiment more with possible performance improvements
of the light engine, I'm trying to keep block access as simple as
possible, so I tried only allowing palettes where the bits per block
is a power of 2. This reduces the register usage and the number of
instructions in SectionGetBlockState.

The performance of the light engine improved a bit, but sadly not
much. Perhaps a better idea is to keep a list of active chunks, that
need their blocks accessed frequently and have the full 16 bits per
block, and a list of inactive chunks that are rarely accessed and are
compressed. We could even compress those inactive chunks using better
compression algorithms than palettes. More experimentation is
definitely in order!

We also deal with some edge cases in the chunk NBT data better now.