Propagate max sky light down first
This has a less random memory access pattern and decreases the number
of block/light reads when we calculate the sky light in some cases.
Another benefit is that we don't need a special propagate-and-push
loop for sky light, since the special downwards propagation of max sky
light has already been handled. This improves the performance of sky
light propagation too, as there are less branches/variables to keep
track off.