Use 1 byte per light value
The good part is that this improves the performance of the light
engine by 6-14% across different kinds of worlds. The bad part is that
this doubles memory usage in void worlds. For now I'm not worrying too
much about the extra memory usage, but we'll definitely have to do
some memory optimisation in the future. There are methods that let us
make significant savings in common scenarios.