Use pext for light engine index calculations
The pext instruction on x86 is very useful for these kinds of
computations. To my surprise, pext speeds up the entire light engine
by around 40%. For now we only actually use pext when compiling for a
CPU that supports BMI2. In the future we will probably do some sort of
runtime detection. Apparently pext is very slow on Zen 2, so our
detection system will have to filter that out.