Tons of 1.21.1 fixes
The update to 1.21.1 is mostly done now, except for item and block
behaviours we need to update and some packet parsing for packets we
don't currently handle anyway. I'll postpone those to a later date
when the behaviour system and the packet system are in a better state
that lends itself better to updates.